[server] 1. remove worker mode and json interceptor, 2.change package to interface

+++ util.go
@@ -4,16 +4,16 @@
 
 package server
 
-const CMD_PER_MOD uint16 = 100
+const CMD_PER_MOD uint32 = 100
 
-func GetMod(protoNo uint16) uint16 {
-	return uint16(protoNo / CMD_PER_MOD)
+func GetMod(protoNo uint32) uint32 {
+	return protoNo / CMD_PER_MOD
 }
 
-func GetCmd(protoNo uint16) uint16 {
-	return uint16(protoNo % CMD_PER_MOD)
+func GetCmd(protoNo uint32) uint32 {
+	return protoNo % CMD_PER_MOD
 }
 
-func GetProtoNo(mod uint16, cmd uint16) uint16 {
-	return uint16(mod*CMD_PER_MOD + cmd)
+func GetProtoNo(mod uint32, cmd uint32) uint32 {
+	return mod*CMD_PER_MOD + cmd
 }
